// apps/web/src/client/presentation/classic-creature-definitions.ts
export const classicCreatureDefinitions = [
  {
    kind: 'pig',
    name: '猪',
    modelId: 'seedlands:model/actor/pig',
    path: 'models/classic/pig.glb',
    byteLength: 21024,
    nodeCount: 17,
    triangleCount: 96,
    height: 0.875,
    clips: {
      idle: 'idle',
      move: 'move',
      attack: 'attack',
      hurt: 'hurt',
    },
  },
  {
    kind: 'cow',
    name: '牛',
    modelId: 'seedlands:model/actor/cow',
    path: 'models/classic/cow.glb',
    byteLength: 22980,
    nodeCount: 21,
    triangleCount: 120,
    height: 1.5,
    clips: {
      idle: 'idle',
      move: 'move',
      attack: 'attack',
      hurt: 'hurt',
    },
  },
  {
    kind: 'sheep',
    name: '羊',
    modelId: 'seedlands:model/actor/sheep',
    path: 'models/classic/sheep.glb',
    byteLength: 16992,
    nodeCount: 13,
    triangleCount: 72,
    height: 1.09375,
    clips: {
      idle: 'idle',
      move: 'move',
      attack: 'attack',
      hurt: 'hurt',
    },
  },
  {
    kind: 'chicken',
    name: '鸡',
    modelId: 'seedlands:model/actor/chicken',
    path: 'models/classic/chicken.glb',
    byteLength: 22656,
    nodeCount: 21,
    triangleCount: 120,
    height: 0.9375,
    clips: {
      idle: 'idle',
      move: 'move',
      attack: 'attack',
      hurt: 'hurt',
    },
  },
  {
    kind: 'squid',
    name: '鱿鱼',
    modelId: 'seedlands:model/actor/squid',
    path: 'models/classic/squid.glb',
    byteLength: 27052,
    nodeCount: 19,
    triangleCount: 108,
    height: 1.1875,
    clips: {
      idle: 'idle',
      move: 'move',
      attack: 'attack',
      hurt: 'hurt',
    },
  },
  {
    kind: 'wolf',
    name: '狼',
    modelId: 'seedlands:model/actor/wolf',
    path: 'models/classic/wolf.glb',
    byteLength: 23936,
    nodeCount: 21,
    triangleCount: 120,
    height: 1.03125,
    clips: {
      idle: 'idle',
      move: 'move',
      attack: 'attack',
      hurt: 'hurt',
    },
  },
  {
    kind: 'zombie',
    name: '僵尸',
    modelId: 'seedlands:model/actor/zombie',
    path: 'models/classic/zombie.glb',
    byteLength: 16812,
    nodeCount: 13,
    triangleCount: 72,
    height: 2,
    clips: {
      idle: 'idle',
      move: 'move',
      attack: 'attack',
      hurt: 'hurt',
    },
  },
  {
    kind: 'skeleton',
    name: '骷髅',
    modelId: 'seedlands:model/actor/skeleton',
    path: 'models/classic/skeleton.glb',
    byteLength: 35092,
    nodeCount: 35,
    triangleCount: 204,
    height: 2,
    clips: {
      idle: 'idle',
      move: 'move',
      attack: 'attack',
      hurt: 'hurt',
    },
  },
  {
    kind: 'spider',
    name: '蜘蛛',
    modelId: 'seedlands:model/actor/spider',
    path: 'models/classic/spider.glb',
    byteLength: 43548,
    nodeCount: 39,
    triangleCount: 228,
    height: 0.625,
    clips: {
      idle: 'idle',
      move: 'move',
      attack: 'attack',
      hurt: 'hurt',
    },
  },
  {
    kind: 'creeper',
    name: '爬行者',
    modelId: 'seedlands:model/actor/creeper',
    path: 'models/classic/creeper.glb',
    byteLength: 17928,
    nodeCount: 13,
    triangleCount: 72,
    height: 1.75,
    clips: {
      idle: 'idle',
      move: 'move',
      attack: 'attack',
      hurt: 'hurt',
    },
  },
  {
    kind: 'slime',
    name: '史莱姆',
    modelId: 'seedlands:model/actor/slime',
    path: 'models/classic/slime.glb',
    byteLength: 11848,
    nodeCount: 11,
    triangleCount: 60,
    height: 1,
    clips: {
      idle: 'idle',
      move: 'move',
      attack: 'attack',
      hurt: 'hurt',
    },
  },
  {
    kind: 'pig-zombie',
    name: '僵尸猪人',
    modelId: 'seedlands:model/actor/pig-zombie',
    path: 'models/classic/pig-zombie.glb',
    byteLength: 21392,
    nodeCount: 19,
    triangleCount: 108,
    height: 2,
    clips: {
      idle: 'idle',
      move: 'move',
      attack: 'attack',
      hurt: 'hurt',
    },
  },
] as const;